Improved model check times for Cisco models (#147)
The time taken to run CI checks on Cisco native models has been lengthening such that it can sometimes take 45 minutes or more to perform all the checks. This has become a problem for all users.

This commit implements parallelisation of the various builds down to the per-OS release level, allowing more CPU resources to be dedicated to model checking as it has become increasingly clear that the serialised pyang invocations do not make best use of available resources.