pdating script according to new yang catalog
Populating of yangdatabase is faster and is done with two requests
rather than with a request per each yang file. Prepare script will
create a json file that will be able to populate the database

yangIntegrity script is made for Einar to be able to see which files
and submodules are missing which are extra and also he will see missing
or wrong namespace, revision.

runCapabilites and statistics script creates some basic statistics but
this is still to be done together with statisticsInCatalog scritp.
We need to pull the statistics code from runCapabilites and move it to
statisticsInCatalog.

Script will also check if there was any change on capability.xml file
and if not we will not parse that folder.

Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>