Add methods for indexing plus start tree-type
New api methods PUT and POST for populating modules
and vendors. This will provide data for indexing and
it will be either all data need to be reindexed or only
the one that are not currently in yang catalog.

Start new metadata parsing - tree type - this is not
done yet.

Cleaning up code a little.

Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>