Automatic push of ietf modules
This will check for new ietf draft modules
and it will automaticaly add them to the
repository. Merging is ready but commented
out to test it first.

This also checks for new RFC modules. If
there are any it will send a notification
via E-mail. (Work in progress for spark
message)

Also some bugfixing

Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>