Setup cisco spark automatic messaging
For several situation we need notification about them.
For couple of them it was set to send email but that
would not be sufficient if we want to send less important
messages. Here we introduces messaging through
messageFactory class which sends email and/or spark
messages to yang catalog admin room.

Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>