Major changes described bellow
Change the way of searching for relative paths. It
now finds the path of a file the script is written
in and then it will add relative path to what is
looked for

Fix Repoutil for adding deleted files.

Fix pull and merge of everyday update of ietf yang
files

Add script in sandbox to remove some unnecessary
yang fies

Create a generic cisco spark message for any
automated tool.

Fix bugs

Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>