Remove experimental YANG for LLDP, in lieu of draft YANG from IEEE P802.1ABcu.
I uploaded this experimental YANG for LLDP before there was a formal project in the 802.1 working group.
Now that 802.1 has a project P802.1ABcu to standardize YANG for LLDP:
  https://development.standards.ieee.org/get-file/P802.1ABcu.pdf?t=94648000003
and a draft of that module exists in:
  https://github.com/YangModels/yang/blob/master/standard/ieee/802.1/draft/ieee802-dot1ab-lldp.yang
there is no longer a need for this experimental module to exist.
Leaving it in experimental might cause uneccesary confusion as to the correct module for LLDP.