Remove experimental YANG for 802.1Qbv, in lieu of draft YANG for 802.1Qbv.
I uploaded this experimental YANG for 802.1Qbv (scheduled traffic) before there was a formal project in the 802.1 working group.
Now that 802.1 has a project P802.1Qcw to standardize YANG for scheduled traffic, and a draft of that module exists in:
   https://github.com/YangModels/yang/blob/master/standard/ieee/802.1/draft/ieee802-dot1q-sched.yang
there is no longer a need for this experimental module to exist.
Leaving it in experimental might cause unnecessary confusion as to the correct module for scheduled traffic.