Add experimental module for IEEE Std 1588 revision.
The upcoming revision of IEEE Std 1588 is starting IEEE Sponsor Ballot in September 2018.
Although YANG is not a formal part of that revision, the content is sufficiently stable to allow for a YANG contribution.
Although this experimental module is intended as input to a future IEEE 1588 PAR, it is not incorporated into IEEE procedures as yet, and therefore it uses a vendor-specific module name and namespace.