Add experimental module for IEEE Std 802.1AS revision.
The upcoming revision of IEEE Std 802.1AS is in IEEE 802.1 Working Group Recirculation Ballot.
Although YANG is not a formal part of that revision, the content is sufficiently stable to allow for a YANG contribution.
Although this experimental module is intended as input to a future IEEE 802.1 PAR, it is not incorporated into IEEE procedures as yet, and therefore it uses a vendor-specific module name and namespace.
Since IEEE Std 802.1AS is a profile of IEEE Std 1588, this module is implemented as an augment to the module for the IEEE Std 1588 revision.